Adjust the order of task groups in an OKR.
The default order is alphabetic: done, in progress, to do. This patch
did a reverse on the order, so the new order will be like this:
To do, in progress, done.

This patch also did a tiny adjustment of font style, cancelled the bold
style for the task groups (to do, in progress, to do) to help make the
whole slide more reasonable.